chore: merge funding mod into hedging mod
