--- funding/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=postgres://stablesats:stablesats@127.0.0.1:5433/stablesats-hedging
