--- funding/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "funding"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-
-fail-on-warnings = []
-
-[dependencies]
-shared = { path = "../shared", package = "stablesats-shared" }
-okex-client = { path = "../okex-client" }
-galoy-client = { path = "../galoy-client" }
-# hedging = { path = "../hedging"}
-
-rust_decimal_macros = "1.26.1"
-serde = "1.0.145"
-serde_json = "1.0.86"
-tokio = "1.21.2"
-futures = "0.3.24"
-thiserror = "1.0.37"
-opentelemetry = { version = "0.18.0", features = ["trace"] }
-tracing-opentelemetry = "0.18.0"
-tracing = "0.1.37"
-chrono = { version = "0.4.22", features = ["clock", "serde"], default-features = false }
-sqlx = { version = "0.6", features = [ "offline", "runtime-tokio-rustls", "postgres", "decimal", "uuid", "chrono"] }
-sqlxmq = { version = "0.4.1", default-features = false, features = [ "runtime-tokio-rustls" ] }
-rust_decimal = "1.26.1"
-uuid = "1.1.2"
-serde_with = "2.0.1"
-
-[dev-dependencies]
-anyhow = "1.0.65"
-serial_test = "*"
