--- funding/src/lib.rs
@@ -1,34 +1,0 @@
-#![cfg_attr(feature = "fail-on-warnings", deny(warnings))]
-#![cfg_attr(feature = "fail-on-warnings", deny(clippy::all))]
-
-mod app;
-mod error;
-mod job;
-mod okex_transfers;
-mod rebalance_action;
-mod synth_usd_liability;
-
-use galoy_client::GaloyClientConfig;
-use okex_client::OkexClientConfig;
-use shared::{health::HealthCheckTrigger, pubsub::*};
-
-pub use app::*;
-pub use error::*;
-
-pub async fn run(
-    health_check_trigger: HealthCheckTrigger,
-    config: FundingAppConfig,
-    okex_config: OkexClientConfig,
-    galoy_config: GaloyClientConfig,
-    pubsub_cfg: PubSubConfig,
-) -> Result<(), FundingError> {
-    FundingApp::run(
-        health_check_trigger,
-        config,
-        okex_config,
-        galoy_config,
-        pubsub_cfg,
-    )
-    .await?;
-    Ok(())
-}
