+++ hedging/src/job/poll_okex.rs
@@ -27,6 +27,19 @@ pub async fn execute(
         })
         .await?;
 
+    // let PositionSize {
+    //     usd_cents,
+    //     instrument_id,
+    //     ..
+    // } = okex.get_position_in_signed_usd_cents().await?;
+    // publisher
+    //     .publish(OkexBtcUsdSwapPositionPayload {
+    //         exchange: ExchangeIdRaw::from(OKEX_EXCHANGE_ID),
+    //         instrument_id: InstrumentIdRaw::from(instrument_id.to_string()),
+    //         signed_usd_exposure: SyntheticCentExposure::from(usd_cents),
+    //     })
+    //     .await?;
+
     let mut execute_sweep = false;
     for id in okex_orders.open_orders().await? {
         match okex.order_details(id.clone()).await {
